Fix build issue of dynamic client updation patch on Android-P
While building the Configure client dynamically without restarting the
AIC server patch on Android-P, build issue was seen because of type
mismatch of cameraId iterator and return type of size of map.

While iterating through size of map using an int type iterator,
it is required to typecast the size of map to int type.

Tracked-On: OAM-104587
Change-Id: I2cf2e50a4ee2a602cdfe6e10982392eaf18ad929
Signed-off-by: Singh, Sapna1 sapna1.singh@intel.com