Fix for camera crash seen after clicking back button on overlay
When the camera is opened and back button of app is pressed,
client disconnects. On relaunching the app, client connects
back and renegotiation takes place that clears the previous
camera info even when the camera session is active resulting
in camera vhal crash.

When camera is in opened state or in decoding state, don't do
renegotiation.

NOTE: This change may need to be reverted once the requirement of
dynamic client configuration when the camera session is active is
clear.

Change-Id: Icff3ca5db8a2c34732885694d620f51dd22d96a1
Tracked-On: OAM-104869
Signed-off-by: Singh, Sapna1 <sapna1.singh@intel.com>