Fix for black camera preview observed in AIC app for H265 config
When camera is launched using the AIC-Client application, there is no
preview and a black screen is seen.

once the camera-vhal receives the encoded frame from client, it will
decode the frame but In this case, after receiving the frame from
client, decoding the frame is failing with error code -14 that is
MFX_ERR_INCOMPATIBLE_VIDEO_PARAM error.

Decoder param codeclevel was not set correctly, required codeclevel
for H265 is 51.

Tracked-on: OAM-105522
Signed-off-by: Singh, Sapna1 <sapna1.singh@intel.com>