Fix crash seen in deleting VirtualCameraFactory object
ConnectionsListener class was derived from Thread class
which was causing issue while calling join for this thread
from VirtualCameraFactory destructor.

The Base class from which ConnectionsListener was derived
was not the proper class.

With current implementation, instead of deriving from thread
class, a thread is introduced as a member of the class to do
the same job and it is started during creation of
ConnectionsListener object

Tracked-On: OAM-106781
Signed-off-by: Singh, Sapna1 <sapna1.singh@intel.com>