Fix for UI not coming when page reloaded with camera app open
Whenever camera is opened and URL is reloaded, camera close will be
called from framework and on the other side streamer will send
capability request to the camera-vhal. camera vhal ignores the
capability request as it is received when the camera is in opened
state. Streamer is stuck waiting for the response to capability
request which results in no UI streamed out.

Fix the issue by handling capability request even when camera 
is in opened state in camera vhal.

Tracked-On: OAM-108772
Signed-off-by: Singh, Sapna1 <sapna1.singh@intel.com>