Fix for crop issue seen when capturing jpeg image
The Captured images in 1920 X 1080 2.1 MP are cropped.

When the images are JPEG Compressed and provided to framework data
more than 300kB is not updated in the buffer resulting in cropped
image. This is because the max JPEG image size is limited to 300kB.

The max size of the JPEG Image is doubled to 600kB to prevent cropping.

Tracked-On: OAM-109712
Signed-off-by: Pillai, Venkatesh <venkatesh.pillai@intel.com>