Fix the unlock and release logic for sensor buffers
The unlock and release of previous sensor buffers were being
attempted using the importedHandle of the current buffer.
Also, the OK check for res along with bLocked flag was causing
the unlock code to be unreachable.

Fixed unlock and release of previous buffers during cleanup.
Removed the bLocked flag, as it is obvious that no lock was
acquired when hitting the cleanup section.

Tracked-On: OAM-111000
Signed-off-by: Anoob Anto K <anoob.anto.kodankandath@intel.com>